fix: Remove job-status
