+++ zeppelin-web/src/app/jobmanager/jobmanager.controller.js
@@ -12,7 +12,14 @@
  * limitations under the License.
  */
 
-import { JobStatus, } from './jobs/job-status'
+export const JobStatus = {
+  READY: 'READY',
+  FINISHED: 'FINISHED',
+  ABORT: 'ABORT',
+  ERROR: 'ERROR',
+  PENDING: 'PENDING',
+  RUNNING: 'RUNNING',
+}
 
 angular.module('zeppelinWebApp')
   .controller('JobManagerCtrl', JobManagerCtrl)
